edited WeatherInfo enabled to assign the data from openweathermap

+++ app/src/main/java/com/kasai/speed_whether/model/WeatherInfo.kt
@@ -7,51 +7,51 @@ data class WeatherInfo(
     val lat: Double,
     val lon: Double,
     val timezone: String,
-    val timezone_offset: Int
+    val timezone_offset: Double
 ) {
     data class Current(
-        val clouds: Int,
+        val clouds: Double,
         val dew_point: Double,
-        val dt: Int,
+        val dt: Double,
         val feels_like: Double,
-        val humidity: Int,
-        val pressure: Int,
-        val sunrise: Int,
-        val sunset: Int,
+        val humidity: Double,
+        val pressure: Double,
+        val sunrise: Double,
+        val sunset: Double,
         val temp: Double,
-        val uvi: Int,
-        val visibility: Int,
+        val uvi: Double,
+        val visibility: Double,
         val weather: List<Weather>,
-        val wind_deg: Int,
+        val wind_deg: Double,
         val wind_gust: Double,
         val wind_speed: Double
     ) {
         data class Weather(
             val description: String,
             val icon: String,
-            val id: Int,
+            val id: Double,
             val main: String
         )
     }
 
     data class Daily(
-        val clouds: Int,
+        val clouds: Double,
         val dew_point: Double,
-        val dt: Int,
+        val dt: Double,
         val feels_like: FeelsLike,
-        val humidity: Int,
+        val humidity: Double,
         val moon_phase: Double,
-        val moonrise: Int,
-        val moonset: Int,
+        val moonrise: Double,
+        val moonset: Double,
         val pop: Double,
-        val pressure: Int,
+        val pressure: Double,
         val rain: Double,
-        val sunrise: Int,
-        val sunset: Int,
+        val sunrise: Double,
+        val sunset: Double,
         val temp: Temp,
         val uvi: Double,
         val weather: List<Weather>,
-        val wind_deg: Int,
+        val wind_deg: Double,
         val wind_gust: Double,
         val wind_speed: Double
     ) {
@@ -74,25 +74,25 @@ data class WeatherInfo(
         data class Weather(
             val description: String,
             val icon: String,
-            val id: Int,
+            val id: Double,
             val main: String
         )
     }
 
     data class Hourly(
-        val clouds: Int,
+        val clouds: Double,
         val dew_point: Double,
-        val dt: Int,
+        val dt: Double,
         val feels_like: Double,
-        val humidity: Int,
+        val humidity: Double,
         val pop: Double,
-        val pressure: Int,
+        val pressure: Double,
         val rain: Rain,
         val temp: Double,
         val uvi: Double,
-        val visibility: Int,
+        val visibility: Double,
         val weather: List<Weather>,
-        val wind_deg: Int,
+        val wind_deg: Double,
         val wind_gust: Double,
         val wind_speed: Double
     ) {
@@ -103,7 +103,7 @@ data class WeatherInfo(
         data class Weather(
             val description: String,
             val icon: String,
-            val id: Int,
+            val id: Double,
             val main: String
         )
     }
